Decompose backup lifecycle service
Move backup create, transfer, validation, and restore lifecycle
ownership out of Sandbox so lifecycle recovery is isolated from the
Durable Object facade and each backup concern has a focused module.
